Processor moved into separate package
