+++ handlers/holiday-stop-processor/src/main/scala/com/gu/creditprocessor/ProcessResult.scala
@@ -1,4 +1,4 @@
-package com.gu.holidaystopprocessor
+package com.gu.creditprocessor
 
 import com.gu.holiday_stops.{CreditRequest, OverallFailure, ZuoraApiFailure}
 import com.typesafe.scalalogging.LazyLogging
