+++ handlers/holiday-stop-processor/src/main/scala/com/gu/holidaystopprocessor/StandaloneApp.scala
@@ -4,6 +4,7 @@ import java.time.LocalDate
 
 import com.gu.effects.GetFromS3
 import com.gu.holiday_stops.Config
+import com.gu.holidaystopprocessor.HolidayStopCreditProcessor.processAllProducts
 import com.softwaremill.sttp.HttpURLConnectionBackend
 
 // This is just for functional testing locally.
@@ -14,7 +15,7 @@ object StandaloneApp extends App {
   Config(GetFromS3.fetchString) match {
     case Left(msg) => println(s"Config failure: $msg")
     case Right(config) =>
-      val processResult = Processor.processAllProducts(config, stopDate, HttpURLConnectionBackend(), GetFromS3.fetchString)
+      val processResult = processAllProducts(config, stopDate, HttpURLConnectionBackend(), GetFromS3.fetchString)
 
       println(processResult.flatMap(_.creditsToApply).size)
