--- handlers/holiday-stop-processor/src/main/scala/com/gu/holidaystopprocessor/package.scala
@@ -1,2 +1,0 @@
-package com.gu
-
